+++ lib/Models/Cesium.ts
@@ -60,7 +60,7 @@ import PickedFeatures, {
 } from "../Map/PickedFeatures/PickedFeatures";
 import MappableMixin, {
   ImageryParts,
-  isCesium3DTileset,
+  isPrimitive,
   isDataSource,
   isTerrainProvider,
   MapItem
@@ -542,6 +542,7 @@ export default class Cesium extends GlobeOrMap {
   }
 
   private observeModelLayer() {
+    let prevMapItems: MapItem[] = [];
     return autorun(() => {
       // TODO: Look up the type in a map and call the associated function.
       //       That way the supported types of map items is extensible.
@@ -606,27 +607,25 @@ export default class Cesium extends GlobeOrMap {
         }
       }
 
-      const allCesium3DTilesets = this._allMapItems.filter(isCesium3DTileset);
-      // Remove deleted tilesets
+      const allPrimitives = this._allMapItems.filter(isPrimitive);
+      const prevPrimitives = prevMapItems.filter(isPrimitive);
       const primitives = this.scene.primitives;
-      // Iterate backwards because we're removing items.
-      for (let i = this.scene.primitives.length - 1; i >= 0; i--) {
-        const prim = primitives.get(i);
-        if (
-          isCesium3DTileset(prim) &&
-          allCesium3DTilesets.indexOf(prim) === -1
-        ) {
-          this.scene.primitives.remove(prim);
-        }
-      }
 
-      // Add new tilesets
-      allCesium3DTilesets.forEach(tileset => {
-        if (!primitives.contains(tileset)) {
-          primitives.add(tileset);
+      // Remove deleted primitives
+      prevPrimitives.forEach(primitive => {
+        if (!allPrimitives.includes(primitive)) {
+          primitives.remove(primitive);
         }
       });
 
+      // Add new primitives
+      allPrimitives.forEach(primitive => {
+        if (!primitives.contains(primitive)) {
+          primitives.add(primitive);
+        }
+      });
+
+      prevMapItems = this._allMapItems;
       this.notifyRepaintRequired();
     });
   }