+++ CHANGES.md
@@ -8,6 +8,7 @@ Change Log
 * Add WPS support for `Date` (additional to existing `DateTime`) and support for `ComplexData` `Date`/`DateTime` WPS Inputs.
 * TSXified `StandardUserInterface` and some other components. If your TerriaMap imports `StandardUserInterface.jsx` remove the `.jsx` extension so webpack can find the new `.tsx` file.
 * Fix use of `baseMapContrastColor` in region mapping/protomaps and remove `MAX_SELECTABLE_DIMENSION_OPTIONS`
+* `mapItems` can now return arbitrary Cesium primitives.
 * [The next improvement]
 
 #### 8.2.4 - 2022-05-23